@@ -4,13 +4,13 @@ description: Advent of Code
 # homepage: https://github.com/praful/advent_of_code
 
 environment:
-  sdk: ">=2.8.1 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  collection: any
-  trotter: any
-  tuple: any
+  collection: ^1.15.0
+  trotter: ^2.0.0-dev.1
+  tuple: ^2.0.0-nullsafety.0
   # basics: any
 
 dev_dependencies:
-  pedantic: ^1.9.0
+  pedantic: ^1.11.0
