@@ -1,10 +1,10 @@
-import './utils.dart';
+import '../../shared/dart/src/utils.dart';
 
 const bool DEBUG = false;
 
-List TEST_INPUT;
-List TEST_INPUT2;
-List MAIN_INPUT;
+late List TEST_INPUT;
+late List TEST_INPUT2;
+late List MAIN_INPUT;
 
 void runPart1(String name, List input) {
   printHeader(name);
@@ -15,8 +15,8 @@ void runPart2(String name, List input) {
 }
 
 void main(List<String> arguments) {
-  TEST_INPUT = fileAsString('../data/day12-test.txt');
-  MAIN_INPUT = fileAsString('../data/day12.txt');
+  TEST_INPUT = FileUtils.asLines('../data/day12-test.txt');
+  MAIN_INPUT = FileUtils.asLines('../data/day12.txt');
 
   //Answer:
   runPart1('12 test part 1', TEST_INPUT);
