@@ -35,26 +35,26 @@ int part1(String header, List<int> input) {
   printHeader(header);
   // return input.fold(0, (a, b) => a + fuelRequired(b));
   // return input.map((f) => fuelRequired(f)).sum; //alternative to above
-  return input.fold2(fuelRequired); //another alt
+  return input.fold2(fuelRequired) as int; //another alt
 }
 
 int part2(String header, List<int> input) {
   printHeader(header);
 
-  return input.fold2(recursiveFuelRequired); //another alt
+  return input.fold2(recursiveFuelRequired) as int; //another alt
   // return input.fold(0, (a, b) => a + recursiveFuelRequired(b));
   // return input.fold(0, (a, b) => a + loopFuelRequired(b));
 }
 
 void main(List<String> arguments) {
-  List testInput = FileUtils.asInt('../data/day01-test.txt');
-  List mainInput = FileUtils.asInt('../data/day01.txt');
+  var testInput = FileUtils.asInt('../data/day01-test.txt');
+  var mainInput = FileUtils.asInt('../data/day01.txt');
 
   assertEqual(part1('01 test part 1', testInput), 34241);
   assertEqual(part2('01 test part 2', testInput), 51316);
 
   //Answer:3246455
-  print(part1('01 part 1', mainInput));
+  printAndAssert(part1('01 part 1', mainInput),3246455);
   //Answer:4866824
-  print(part2('01 part 2', mainInput));
+  printAndAssert(part2('01 part 2', mainInput),4866824);
 }
