@@ -1,10 +1,6 @@
-import './utils.dart';
+import '../../shared/dart/src/utils.dart';
 
 const bool DEBUG = false;
-
-List TEST_INPUT;
-List TEST_INPUT2;
-List MAIN_INPUT;
 
 class Point {
   final int x;
@@ -92,7 +88,7 @@ class PocketDimension {
           var updatePoint = Point(x, y, z);
           var activeCount = activeNeighbours(updatePoint);
           if (isActive(updatePoint)) {
-            if (activeCount.isBetweenI(2, 3)) {
+            if (activeCount.isBetween(2, 3)) {
               makeActive(newGrid, updatePoint);
             }
           } else {
@@ -119,8 +115,8 @@ void runPart1(String name, List input) {
 }
 
 void main(List<String> arguments) {
-  TEST_INPUT = fileAsString('../data/day17-test.txt');
-  MAIN_INPUT = fileAsString('../data/day17.txt');
+  var TEST_INPUT = FileUtils.asLines('../data/day17-test.txt');
+  var MAIN_INPUT = FileUtils.asLines('../data/day17.txt');
 
   //Answer: 112
   runPart1('17 test part 1', TEST_INPUT);
